test: Capture gtest log output and only print on error
In zcash/zcash#5762 we altered the gtest runner to turn on logging to
stdout instead of dropping the logs, to make figuring out test failures
easier. However, this also meant the logs would be displayed for tests
that succeeded, and it was confusing to see `ERROR` log lines present.

We now have a test-specific initialization method for tracing that uses
synchronous logging instead of a background thread, and only logs to a
file. In the gtests, we initialize this with a new temporary file, and
add a gtest event listener that clears the file at the start of each
test, and then prints its contents if the test fails.

Example test output:

```
[ RUN      ] ChecktransactionTests.CheckVpubNotBothNonzero
gtest/test_checktransaction.cpp:39: Failure
Expected equality of these values:
  state.GetRejectReason()
    Which is: "bad-txns-vpubs-both-nonzero"
  "bad-txns-vpubs-both-nonzer"

--- Logs:
  ERROR main: CheckTransaction(): joinsplit.vpub_new and joinsplit.vpub_old both nonzero
---
[  FAILED  ] ChecktransactionTests.CheckVpubNotBothNonzero (0 ms)
```

Closes zcash/zcash#5766.