wallet: Remove lock on cs_main from CWallet::ChainTip
We must not lock on cs_main during wallet synchronization, to avoid
locking side channels between the wallet and the networking stack.
However, the wallet code that is called on every block to update tree
states etc. was also checking whether it should run Sprout-to-Sapling
migrations, and that was being gated on IBD completion (to avoid stale
transactions being generated). IsInitialBlockDownload locks on cs_main
before IBD completion :)

We now rely on the fact that IBD does not end until the chain tip is
within nMaxTipAge of the current time, so we can check that as a proxy.
We were already applying a similar check to avoid hibernation-related
issues; these are now merged.