Check dependency updates on the correct branch
Previously make-release.py checked dependency updates before checking out the
commit to be released. This could be annoying if the release branch had
everything up-to-date, but you ran the script from a different branch. Worse,
you could accidentally update the dependencies on your original branch, and have
it pass the release process with outdated dependencies.