Encapsulate memos
There’s a `Memo` class that was introduced with the WalletTxBuilder work, but it
was restricted to the wallet. This moves it to libzcash, uses it universally,
and makes it a lot more robust.

The only place (outside of tests) that uses `Memo::Bytes` is
`BaseNotePlaintext`, which needs a particular in-memory representation. But the
API for `*NotePlaintext` is in terms of the `Memo` class.