Avoid extra copy in ResolveInputsAndPayments
This change takes `const SpendableInputs&` instead of copying it. It does an explicit copy later in
more limited scope, avoiding an extra copy in the conventional fee side.

However, this means both the reference and mutable copy are in scope at the same time, making it
easy to use the wrong one by mistake.