Enable mining to Orchard on regtest
The `HeightOfLatestSettledUpgrade` function depends upon the existence
of checkpoints for its determination of current chain height that is
used in the selection of the correct receiver for a unified miner
address. Since we never create checkpoints on regtest, this was
previously preventing regtest nodes from mining to Orchard shielded
coinbase because NU5 was not considered active by `GetPreferredRecipientAddress`
given the height returned by `HeightOfLatestSettledUpgrade`.