Fix Orchard bridgetree parsing.
Prior to v5.6.0, `BridgeTree` serialization did not write a distinct
version identifier for the individual `MerkleBridge` values being
serialized. In v5.6.0, the serialization format was updated to include
such an identifier, but the parsing code for these values erroneously
attempted to read this code when parsing pre-v5.6.0 tree data. When
parsing failed, this caused the Orchard note commitment tree history to
be dropped, which will necessitate a re-scan to repopulate this data in
affected wallets.

While the `wallet_golden_5_6_0` test had been specifically created to
detect this kind of parsing failure, a bug in the test setup meant that
the serialized form which needed to be checked was not correctly
persisted in the way that was necessary to actually trigger the error.