Use at least the ZIP 317 fee for Sprout->Sapling migration.
Because of excessive complexity in the Sprout JoinSplit construction
algorithm, we might not be able to compute the exact ZIP 317 fee, since
we essentially need to predict how many JoinSplits the algorithm will
use. So, we use a conservative upper bound to compute the fee. The
difference betweeen that upper bound and the actual number of JoinSplits
might leak some information.

Signed-off-by: Daira-Emma Hopwood <daira@jacaranda.org>